merge

+++ tunnel/transport/client.go
@@ -2,24 +2,29 @@ package transport
 
 import (
 	"context"
+	"os"
+	"os/exec"
+	"strconv"
+
 	"github.com/p4gefau1t/trojan-go/common"
 	"github.com/p4gefau1t/trojan-go/config"
 	"github.com/p4gefau1t/trojan-go/log"
 	"github.com/p4gefau1t/trojan-go/tunnel"
-	"net"
-	"os"
-	"os/exec"
-	"strconv"
+	"github.com/p4gefau1t/trojan-go/tunnel/freedom"
 )
 
 // Client implements tunnel.Client
 type Client struct {
 	serverAddress *tunnel.Address
 	cmd           *exec.Cmd
+	ctx           context.Context
+	cancel        context.CancelFunc
+	direct        *freedom.Client
 }
 
 func (c *Client) Close() error {
-	if c.cmd != nil {
+	c.cancel()
+	if c.cmd != nil && c.cmd.Process != nil {
 		c.cmd.Process.Kill()
 	}
 	return nil
@@ -31,9 +36,9 @@ func (c *Client) DialPacket(tunnel.Tunnel) (tunnel.PacketConn, error) {
 
 // DialConn implements tunnel.Client. It will ignore the params and directly dial to the remote server
 func (c *Client) DialConn(*tunnel.Address, tunnel.Tunnel) (tunnel.Conn, error) {
-	conn, err := net.Dial("tcp", c.serverAddress.String())
+	conn, err := c.direct.DialConn(c.serverAddress, nil)
 	if err != nil {
-		return nil, common.NewError("transport failed to connect to remote server")
+		return nil, common.NewError("transport failed to connect to remote server").Base(err)
 	}
 	return &Conn{
 		Conn: conn,
@@ -41,7 +46,7 @@ func (c *Client) DialConn(*tunnel.Address, tunnel.Tunnel) (tunnel.Conn, error) {
 }
 
 // NewClient creates a transport layer client
-func NewClient(ctx context.Context, c tunnel.Client) (*Client, error) {
+func NewClient(ctx context.Context, _ tunnel.Client) (*Client, error) {
 	cfg := config.FromContext(ctx, Name).(*Config)
 
 	var cmd *exec.Cmd
@@ -59,7 +64,7 @@ func NewClient(ctx context.Context, c tunnel.Client) (*Client, error) {
 				"SS_LOCAL_PORT="+strconv.FormatInt(int64(pluginPort), 10),
 				"SS_REMOTE_HOST="+cfg.RemoteHost,
 				"SS_REMOTE_PORT="+strconv.FormatInt(int64(cfg.RemotePort), 10),
-				"SS_PLUGIN_OPTIONS="+cfg.TransportPlugin.PluginOption,
+				"SS_PLUGIN_OPTIONS="+cfg.TransportPlugin.Option,
 			)
 			cfg.RemoteHost = pluginHost
 			cfg.RemotePort = pluginPort
@@ -84,9 +89,16 @@ func NewClient(ctx context.Context, c tunnel.Client) (*Client, error) {
 			return nil, common.NewError("invalid plugin type: " + cfg.TransportPlugin.Type)
 		}
 	}
+
+	direct, err := freedom.NewClient(ctx, nil)
+	common.Must(err)
+	ctx, cancel := context.WithCancel(ctx)
 	client := &Client{
 		serverAddress: serverAddress,
 		cmd:           cmd,
+		ctx:           ctx,
+		cancel:        cancel,
+		direct:        direct,
 	}
 	return client, nil
 }
